// src/modules/users/enums/user-operation-summaries.enum.ts
/* src/modules/users/enums/user-operation-summaries.enum.ts */
export enum UserOperationSummaries {
  CREATE = 'Crear nuevo usuario',
  FIND_ALL = 'Obtener todos los usuarios',
  FIND_ONE = 'Obtener usuario por ID',
  FIND_BY_PHONE = 'Obtener usuario por teléfono',
  UPDATE = 'Actualizar usuario',
  DELETE = 'Eliminar usuario',
  ADD_PARTICIPATION = 'Agregar participación a usuario',
  REMOVE_PARTICIPATION = 'Remover participación de usuario',
}