// src/modules/users/enums/user-messages.enum.ts
/* src/modules/users/enums/user-messages.enum.ts */
export enum UserMessages {
  USER_CREATED = 'Usuario creado exitosamente',
  USER_NOT_FOUND = 'Usuario no encontrado',
  USERS_LISTED = 'Listado de usuarios',
  USER_UPDATED = 'Usuario actualizado exitosamente',
  USER_DELETED = 'Usuario eliminado exitosamente',
  PHONE_CONFLICT = 'El teléfono ya está en uso',
  PARTICIPATION_ADDED = 'Participación añadida al usuario',
  PARTICIPATION_REMOVED = 'Participación removida del usuario',
  INVALID_ID = 'ID de usuario inválido',
}